==> InsuranceRegression/experimentation/00_baseline.ipynb <==
# General Baseline

Purpose: We want to obtain the simplest baseline possible. In this case, something like always returning the average of all Y values or something with using a single column may be close to a decent score and may be all that is needed in the business world

Result:
Using the median value on the raw dataset places me in about the top 55% of the competition with RMSE of 1.108!

%pip install -r ../../requirements.txt

from utilities.kaggle_utils import *

competition_name = "playground-series-s4e12"
download(competition_name)

#load data
train_df = pd.read_csv('../data/raw_dataset/train.csv')
test_df = pd.read_csv('../data/raw_dataset/test.csv')

#get the mean and median
y_col = 'Premium Amount'
median = train_df[y_col].median()
mean = train_df[y_col].mean()

# also get the clipped mean
lower = train_df[y_col].quantile(0.01)
upper = train_df[y_col].quantile(0.99)
train_df[y_col] = train_df[y_col].clip(lower, upper)

clipped_mean = train_df[y_col].mean()

# get the needed ids and insert the mean and median
base_results = test_df[['id']]
mean_results = base_results.copy()
median_results = base_results.copy()
clipped_mean_results = base_results.copy()

mean_results[y_col]=mean
median_results[y_col]=median
clipped_mean_results[y_col] = clipped_mean

result_pth = '../results/'
result_pth_mean = os.path.join(result_pth,'mean_results.csv')
result_pth_median = os.path.join(result_pth,'median_results.csv')
result_pth_clipped_mean = os.path.join(result_pth,'clipped_mean_results.csv')
mean_results.to_csv(result_pth_mean,index=False)
median_results.to_csv(result_pth_median, index=False)
clipped_mean_results.to_csv(result_pth_clipped_mean,index=False)

submit("median_results.csv", "base submission - median")

submit("mean_results.csv", 'base submission - mean')

submit("clipped_mean_results.csv", 'base submission - clippped mean')
