==> InsuranceRegression/experimentation/01_init_EDA.ipynb <==
# Init EDA

purpose of NB: try to understand the data itself

Determine a plan of attack for a baseline approach + next steps on how to clean the dataset and then get into feature engineering


%pip install -r ../requirements.txt

# import/load data


import pandas as pd

train_df = pd.read_csv("../data/raw_dataset/train.csv")

# Save base dataset
This serves as the main dataset used to build the future refined versions of our datasets. Below are the basic changes made:
- drop id col
- drop duplicates
- convert `Policy Start Date` to `Policy Duration Mins`
- convert string data to categorical type


#remove id col and duplicates
train_df = train_df.drop(columns=['id'])
train_df=train_df.drop_duplicates()
#convert start date to a timespan in mins
train_df['Policy Start Date'] = pd.to_datetime(train_df['Policy Start Date'])
train_df['Policy Duration Mins'] = ((pd.Timestamp.now() - train_df['Policy Start Date']).dt.total_seconds())/60
train_df = train_df.drop(columns=['Policy Start Date'])


train_df.to_csv('../data/train_base.csv', index=False)

train_df

# General data info

This section contains insights with shape, null counts, data types

I see many rows of data, which may make model training on full dataset time costly. Will likely want to work on a statistically proportional subset of the dataset for the first few iterations of training

train_df.shape

Here's the breakdown of datatypes:

1 continuous feature (health score)

9 discrete features (8 float types and 1 object -- technically start date is `object` but can be converted to int)

10 categorical


about half of the columns have empty missing data
- this will be addressed on a row-by-row basis

train_df.head(1)

train_df.info()

train_df.describe()

train_df.isnull().sum()

# Plan of attack

0. save the slightly altered dataset (`base_train.csv`)

###  Baseline model-based approach
1. make a baseline dataset (no processing)
    1. remove rows with nulls    
2. Use a few baseline models (no hyperparam tuning)
    1. Linear Regression
    2. Random Forest
    3. SVM
3. Compare against best heuristic approach (always predict median insurance value)
