==> InsuranceRegression/experimentation/02_baseline_dataset.ipynb <==
# Baseline Model

Purpose:

 While we have a minimal understanding of the data, I will now apply the minimal number of preprocessing and train a simple linear regression to get a simple dataset+ model baseline

%pip install -r ../../requirements.txt

# load data

base_dataset = pd.read_csv('../data/train_base.csv')

# Prepare Dataset (minimal preprocessing)

- drop rows with any null columns
- apply label encoding

1.2 mil rows -> 320k
Surprisingly the columwise data distributions remained the same, unclear/(didn't check) if the feature relationships are still in tact

# Drop rows with any null values
base_dataset = base_dataset.dropna()
print(base_dataset.shape)
# Verify if there are any null values left
print(f"are there anymore nulls? {base_dataset.isnull().sum().any()}")  # Should print 0 for all columns


# convert string cols to categorical for future encoding
categorical_cols = [base_dataset.columns[i] for i, x in enumerate(base_dataset.dtypes) if x == 'object']
base_dataset[categorical_cols] = base_dataset[categorical_cols].astype('category')

#make copies for the encodings
base_dataset_label_encoded = base_dataset.copy(deep=True)
base_dataset_one_hot = base_dataset.copy(deep=True)


# Convert categorical to label encodings
for col in categorical_cols:
    base_dataset_label_encoded[col] = base_dataset_label_encoded[col].cat.codes
base_dataset_label_encoded.to_csv("../data/baseline_dataset.csv",index=False)
base_dataset_label_encoded.head()

## convert categorical to one hot encodings
base_dataset_one_hot = pd.get_dummies(base_dataset_one_hot, drop_first=True)
base_dataset_one_hot.to_csv("../data/baseline_dataset_one_hot.csv",index=False)
base_dataset_one_hot.head()
