==> InsuranceRegression/experimentation/04_data_imputation.ipynb <==
# Data imputation

Purpose:
The goal of the Notebook is to handle any nulls in the columns before preprocessing

base_dataset = pd.read_csv('../data/01_train_base.csv')

#since we found one-hot encoding better than label encoding, we'll apply that first
# convert string cols to categorical
categorical_cols = [base_dataset.columns[i] for i, x in enumerate(base_dataset.dtypes) if x == 'object']
base_dataset[categorical_cols] = base_dataset[categorical_cols].astype('category')

## convert categorical to one hot encodings
base_dataset = pd.get_dummies(base_dataset, drop_first=True)
base_dataset.head(1)

# imputations

base_dataset_path = "../data/"
dataset_version = "03_"
y_col = "Premium Amount"

## 1. simple median imputation

simple_median_dataset = base_dataset.copy()

simple_median_dataset = simple_median_dataset.fillna(simple_median_dataset.median())

file_name = dataset_version+ "imputation_median.csv"
save_path = os.path.join(base_dataset_path,file_name)

simple_median_dataset.to_csv(save_path,index=False)


## 2. Mean imputation (after clipping)

For now we'll use quantile clipping instead of z-score since some colmns are skewed

clipped_mean_dataset = base_dataset.copy()


# Apply clipping column-wise
for column in clipped_mean_dataset.columns:
    if clipped_mean_dataset[column].dtype in ['float64', 'int64'] and column != y_col:  # Only process numeric columns
        lower = clipped_mean_dataset[column].quantile(0.01)
        upper = clipped_mean_dataset[column].quantile(0.99)
        clipped_mean_dataset[column] = clipped_mean_dataset[column].clip(lower, upper)

clipped_mean_dataset = clipped_mean_dataset.fillna(clipped_mean_dataset.mean())


file_name = dataset_version+ "imputation_clipped_mean.csv"
save_path = os.path.join(base_dataset_path,file_name)

clipped_mean_dataset.to_csv(save_path,index=False)

## 3. median/mean imputation (cleared outliers + grouping-derived values)
group derived values in this case might be where y value is within a certain range (ie 10 bins), then calculate the mean/median
